pyproject.toml
[tool.poetry]
name = "spor"
version = "1.1.3"
description = 'A system for anchoring metadata in external files to source code'
authors = ["Austin Bingham <austin@sixty-north.com>"]
url = 'https://github.com/abingham/spor'
repository = 'https://github.com/abingham/spor'
license = "MIT"
readme = 'README.rst'

packages = [
    { include = "spor", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.4"
docopt_subcommands = "^2.3"
exit_codes = "^1.1"
pyyaml = "^3.13"

[tool.poetry.dev-dependencies]
hypothesis = "^3.79"
pytest = "^3.9"
radish-bdd = "^0.10.0"
tox = "^3.5"

[tool.poetry.scripts]
spor = "spor.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
